% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{al15}
\alias{al15}
\title{Generate vector of 15 alpha values}
\usage{
al15()
}
\value{
Vector of alpha values
}
\description{
Useful to compute confidence intervals using the `lims` function with P=al15().
}
