% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{al49}
\alias{al49}
\title{Generate Vector of 49 alpha values}
\usage{
al49()
}
\value{
Vector of alpha values
}
\description{
Useful to compute confidence intervals using the `lims` function with P=al49().
}
