% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rt.R
\name{check_rt}
\alias{check_rt}
\title{Check reaction times}
\usage{
check_rt(data, ratio = 1/3)
}
\arguments{
\item{data}{The output from gonogo()}

\item{ratio}{The ratio of the distance between any two (sorted) data points. For example, the default 1/3 means that the user is notified when there is a distance between any two sorted data points that is greater than one third of the range of the data.}
}
\value{
Returns tips for processing the reaction times
}
\description{
Check reaction times
}
\examples{
function(data, ratio = 1/4)
}
