% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim.R
\name{clim}
\alias{clim}
\title{Title}
\usage{
clim(rx, ry, m, s, uu, levb)
}
\arguments{
\item{rx}{}

\item{ry}{}

\item{m}{}

\item{s}{}

\item{uu}{}

\item{levb}{}
}
\description{
Title
}
