% Generated by roxygen2: do not edit by hand
\name{dataoutput}
\alias{dataoutput}
\title{{Output data structure}}
\usage{
  dataoutput(prob.val, output.format = "OC")
}
\arguments{
  \item{prob.val}{Probabilities of success based on different scenarios with list format.}
  \item{output.format}{"OC" (default): Operating characteristics; "PP": Posterior probability.}

}

\value{
  It returns a matrix of output including the inputs and results (Probability of Go and Probability of No-Go or Posterior probability).
}



\description{
  {Output of the result.}
}


\examples{
  start.time = Sys.time()

  n.t = rep(c(8), 6) # Sample size in treatment arm
  mu.t = rep(c(40, 45, 50), 6) # True mean of the treatment arm
  sd.t = rep(15, 6) # True sd of the treatment arm
  hypothesis = "between" # "greater than", "less than", "between"
  threshold =  matrix(rep(c(35, 45), each = 6), ncol = 2)# k: Target region cut-off
  PP.cutoffGo = rep(c(0.6, 0.7), each = 3) # c1: Cut-off for probability of Go
  PP.cutoffNoGo = rep(c(0.6, 0.7), each = 3) # c2: Cut-off for probability of No-Go
  prior = "non-informative" # Prior choice
  reps = 1000 # Number of repetitions for samples from the posterior using simulation
  nsim = 1000 # Number of simulations
  seed.val = 298743 # Seed

  pos.val = vector("list", length(n.t))
  for(i in 1:length(n.t)){
    pos.val[[i]] = Normal1SampleGNG(n.t = n.t[i], mu.t = mu.t[i], sd.t = sd.t[i],
                                    hypothesis = hypothesis, threshold = threshold[i,],
                                    PP.cutoffGo = PP.cutoffGo[i],
                                    PP.cutoffNoGo = PP.cutoffNoGo[i],
                                    nsim = nsim, reps = reps)$pos
  }

  dataoutput(prob.val = pos.val, output.format = "OC")

  Sys.time()- start.time
}
