% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawScreen.R
\name{drawScreen}
\alias{drawScreen}
\title{Draws a screen}
\usage{
drawScreen(txt, cex = 1, col = "black")
}
\arguments{
\item{txt}{A character string specifying the text to be printed on the screen}

\item{cex}{A numeric value specifying the font size (default is 1)}

\item{col}{A character string specifying the color of the text (default is black)}
}
\value{
Draws a screen corresponding to the specified values
}
\description{
Draws a screen
}
