% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{fixw}
\alias{fixw}
\title{Drops the last n entries of a test.}
\usage{
fixw(w, k = 1)
}
\arguments{
\item{w}{list containing results}

\item{k}{number of results to drop}
}
\value{
list containing shortened results
}
\description{
If results are saved in a list V, then z=fixw(V,n) allows you to go back and correct an entry.
}
