% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd0.R
\name{gd0}
\alias{gd0}
\title{Title}
\usage{
gd0(xx, d0, dat0, ID, mu, sig, reso, ln, iseed = -1)
}
\arguments{
\item{xx}{}

\item{d0}{}

\item{dat0}{}

\item{ID}{}

\item{mu}{}

\item{sig}{}

\item{reso}{}

\item{ln}{}

\item{iseed}{}
}
\description{
Title
}

