% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getd0.R
\name{getd0}
\alias{getd0}
\title{Title}
\usage{
getd0(xx, d0, dat0, ID, reso, about, titl, unit, ln, cab = F)
}
\arguments{
\item{xx}{}

\item{d0}{}

\item{dat0}{}

\item{ID}{}

\item{reso}{}

\item{about}{}

\item{titl}{}

\item{unit}{}

\item{ln}{}

\item{cab}{}
}
\description{
Title
}

