% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_utils_bin.R
\name{gn_beta_effect_posts}
\alias{gn_beta_effect_posts}
\title{Get posterior samples}
\usage{
gn_beta_effect_posts(
  n1,
  y1,
  n0 = NULL,
  y0 = NULL,
  phat_ctl = 0.2,
  pri_trt = c(0.25, 0.25),
  pri_ctl = c(0.25, 0.25),
  ...,
  seed = NULL
)
}
\description{
Get posterior samples
}
