% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_utils_bin.R
\name{gn_beta_simu_all}
\alias{gn_beta_simu_all}
\title{Simulate all trials}
\usage{
gn_beta_simu_all(
  vec_n,
  vec_true_trt,
  vec_true_ctl,
  rand_ratio = 0,
  ...,
  seed = NULL,
  n_cores = 4,
  n_simu = 500
)
}
\arguments{
\item{rand_ratio}{randomization ratio. 0 corresponds to single arm}
}
\description{
Simulate all trials
}
