% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_utils_bin.R
\name{gn_beta_simu_comb}
\alias{gn_beta_simu_comb}
\title{Simulate all combinations}
\usage{
gn_beta_simu_comb(vec_n, rand_ratio = 0, phat_ctl = NULL, ...)
}
\arguments{
\item{rand_ratio}{randomization ratio. 0 corresponds to single arm}
}
\description{
Simulate all combinations
}
