% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_utils_bin.R
\name{gn_beta_simu_single}
\alias{gn_beta_simu_single}
\title{Simulate individual trial}
\usage{
gn_beta_simu_single(n1, n0 = NULL, true_trt, true_ctl, ..., seed = NULL)
}
\description{
Simulate individual trial
}
