% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{gn_decision}
\alias{gn_decision}
\title{Go no-go decision}
\usage{
gn_decision(
  post_effect,
  lrv = 0.2,
  tv = 0.05,
  thresh_lrv = 0.8,
  thresh_tv = 0.2,
  rule_go = 1,
  rule_nogo = 2,
  option = c("gofirst", "nogofirst", "consider", "conflict")
)
}
\arguments{
\item{rule_nogo}{1: fail to meet lrv, 2: fail to meet tv, 3: fail to meet
either, 4: fail to meet neither}

\item{rule_g0}{1: meet lrv, 2: meet tv, 3: meet either, 4: meet both}
}
\description{
Go no-go decision
}
