% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{gn_decision_binary_opc}
\alias{gn_decision_binary_opc}
\title{Get Binary decision OPC}
\usage{
gn_decision_binary_opc(dec_tbl, vec_true_rate_trt, true_rate_ctl = 1)
}
\description{
Get Binary decision OPC
}
