% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{gn_decision_opc}
\alias{gn_decision_opc}
\title{Go no-go decision operating characteristics}
\usage{
gn_decision_opc(para_all, post_effect, ...)
}
\description{
Go no-go decision operating characteristics
}
