% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{gn_decision_single}
\alias{gn_decision_single}
\title{Meet a specific threshold}
\usage{
gn_decision_single(
  post_effect,
  cut,
  prob_thresh,
  type_better = c("bigger", "less")
)
}
\description{
Check if the probability of the treatment effect better than the cut value is
more than the probaility threshold.
}
