% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_shiny.R
\name{gn_shiny}
\alias{gn_shiny}
\title{Run Web-Based application}
\usage{
gn_shiny(appname = "shiny", pkgname = "gonogo")
}
\description{
Call Shiny to run \code{statidea} as a web-based application.
}
\details{
A web browser will be brought up for users to access the GUI
}
