% Generated by roxygen2: do not edit by hand
\name{GNGplot}
\alias{GNGplot}
\title{{Plot of probability of successes}}
\usage{
Normal1SampleGNGplot(data = data, x.var = data$mu.t,
                     fill.var = data$PP.cutoffGo,
                     hypothesis = data$hypothesis,
                     x.lab = "mu.t",
                     title.text =  paste("PP(True mean is", hypothesis,
                     "threshold) >= ", PP.cutoffGo*100, "\%"),
                     legend.title.text = "PP.cutoffGo")
}
\arguments{
\item{data}{Output from the model.}

\item{x.var}{Variable for x-axis.}

\item{fill.var}{Variable for fill.}

\item{hypothesis}{Hypothesis.}

\item{x.lab}{Label of x-axis.}

\item{title.text}{Title of the figure.}

\item{legend.title.text}{Legend title of the figure.}
}

\value{
It returns the plot of probability of Go.
}


\details{
Code to calculate the posterior probabilities with probabilities of Go.
}
\description{
{Plot of probability of successes.}
}


\examples{

start.time = Sys.time()

n.t = rep(c(8), 6) # Sample size in treatment arm
mu.t = rep(c(40, 45, 50), 6) # True mean of the treatment arm
sd.t = rep(15, 6) # True sd of the treatment arm
hypothesis = "between" # "greater than", "less than", "between"
threshold =  matrix(rep(c(35, 45), each = 6), ncol = 2)# k: Target region cut-off
PP.cutoffGo = rep(c(0.6, 0.7), each = 3) # c1: Cut-off for probability of Go
PP.cutoffNoGo = rep(c(0.6, 0.7), each = 3) # c2: Cut-off for probability of No-Go
prior = "non-informative" # Prior choice
reps = 1000 # Number of repetitions for samples from the posterior using simulation
nsim = 1000 # Number of simulations

pos.val = vector("list", length(n.t))
for(i in 1:length(n.t)){
  pos.val[[i]] = Normal1SampleGNG(n.t = n.t[i], mu.t = mu.t[i], sd.t = sd.t[i],
                                  hypothesis = hypothesis, threshold = threshold[i,],
                                  PP.cutoffGo = PP.cutoffGo[i],
                                  PP.cutoffNoGo = PP.cutoffNoGo[i],
                                  nsim = nsim, reps = reps)$pos
}

Sys.time()- start.time

data = dataoutput(pos.val = pos.val)

GNGplot(data = data, x.var = data$mu.t,
                     fill.var = data$PP.cutoffGo,
                     hypothesis = data$hypothesis,
                     x.lab = "mu.t",
                     title.text =  paste("PP(True mean is", hypothesis,
                     "threshold) >= PP cutoff"),
                     legend.title.text = "PP.cutoffGo")

GNGplot(data = data, x.var = data$PP.cutoffGo,
                     fill.var = data$mu.t,
                     hypothesis = data$hypothesis,
                     x.lab = "PP.cutoffGo",
                     title.text =  paste("PP(True mean is", hypothesis,
                     "threshold) >= PP cutoff"),
                     legend.title.text = "mu.t" )
}
