% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_package.R
\docType{package}
\name{gonogo-package}
\alias{gonogo-package}
\alias{gonogo}
\title{The 'gonogo' package.}
\description{
Making go or no-go deceisions with Shiny applications.
}
\author{
\strong{Maintainer}: Chenguang Wang \email{chenguang.wang.0517@gmail.com}

}
