% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R, R/main.R
\docType{package}
\name{gonogo}
\alias{gonogo}
\title{This package contains the Go-No Go Task.}
\usage{
gonogo(
  id,
  n_trial = 40,
  n_block = 4,
  stimuli = c("A", "X"),
  inter = 0.6,
  prb = c(0.75, 0.25)
)
}
\arguments{
\item{id}{Either a character vector of length 1 specifying the participant's name, or a numeric vector of length 1 specifying a unique id number}

\item{n_trial}{A numeric vector of length 1 specifying the number of trials per block (default is 40)}

\item{n_block}{A numeric vector of length 1 specifying the number of blocks (default is 4)}

\item{stimuli}{A character vector of length 2 specifying the Go stimulus and the No Go stimulus (in that order; default is A and X)}

\item{inter}{A numeric vector of length 1 specifying the length of each trial in seconds (default is 0.6)}

\item{prb}{A numeric vector of length 2 specifying the probability weights for the Go stimulus and No Go stimulus (in that order; default is 0.75 and 0.25), must add up to 1}
}
\value{
A dataframe consisting of \emph{n_trial*n_block} rows and seven columns: \emph{id} (participant's name or numeric id),  \emph{response} (response key used on the trial),  \emph{correct} (1=correct, 0=incorrect),  \emph{SDT} (responses categorized according to Signal Detection Theory),  \emph{rt} (reaction time in seconds),  \emph{stimulus} (the stimulus shown on the trial), and \emph{block} (the number of the block)
}
\description{
The package has x functions.
}
\section{Validate user's response}{

\emph{validate_user_response} validates the user's response.
}

