% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{gonogoSim}
\alias{gonogoSim}
\title{Run adaptive sensitivity tests in simulation mode.}
\usage{
gonogoSim(
  mlo,
  mhi,
  sg,
  n2 = 0,
  n3 = 0,
  p = 0,
  lam = 0,
  dm = 0,
  ds = 0,
  reso = 0,
  ln = F,
  plt = 0,
  iseed = -1,
  IIgo = T,
  M = 1,
  test = 1,
  BL = c(4, 1, 0)
)
}
\arguments{
\item{mlo}{mu low}

\item{mhi}{mu high}

\item{sg}{sigma guess}

\item{n2}{size of phase II (could be 0)}

\item{n3}{size of phase III (could be 0)}

\item{p}{to approximate the stress level Lp - for phase III only}

\item{lam}{the lambda parameter needed for phase III only}

\item{dm}{deviation from a target mean (tm) based on mlo, mhi, and sg}

\item{ds}{deviation from a target standard deviation (ts) based on mlo, mhi, and sg}

\item{reso}{resolution of stimulus, e.g. 0.01}

\item{ln}{set to T to use log transform which keeps recommended stimulus positive, defaults to F}

\item{plt}{1 (for a history plot of a test) or 0 (no plot, the default)}

\item{iseed}{initialization of the random seed (-1, the default, for full random)}

\item{IIgo}{when T, test proceeds unencumbered; when F, test stops at end of stage 2 of phase I}

\item{M}{factor to scale stresses, defaults to 1}

\item{test}{number corresponding to test type
1 = 3pod
2 = Neyer
3 = Bruceton
4 = Langlie}

\item{BL}{vector of 3 integers, needed for Bruceton or Langlie tests only}
}
\value{
List containing results.
}
\description{
Run adaptive sensitivity tests in simulation mode.
}
