% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxr.R
\name{gxr}
\alias{gxr}
\title{Title}
\usage{
gxr(x, mu, sig, reso, ln, iset)
}
\arguments{
\item{x}{}

\item{mu}{}

\item{sig}{}

\item{reso}{}

\item{ln}{}

\item{iset}{}
}
\description{
Title
}
