% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlrcb.R
\name{jlrcb}
\alias{jlrcb}
\title{Joint & Individual likelihood ratio confidence bounds}
\usage{
jlrcb(dat)
}
\arguments{
\item{dat}{}
}
\description{
Joint & Individual likelihood ratio confidence bounds
}
