% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{lims}
\alias{lims}
\title{Compute confidence limits about probabilities and quantiles}
\usage{
lims(ctyp, dat, conf, P = numeric(0), Q = numeric(0))
}
\arguments{
\item{ctyp}{an integer: 1 for FM; 2 for LR; and 3 for GLM.}

\item{dat}{The d0 component of a named list generated by gonogo (or gonogoSim).
For example, for the previously defined list wWT, dat would be wWT$d0.}

\item{conf}{a confidence associated with the interval, e.g., .95. Think of conf as 2-sided}

\item{P}{a vector of probabilities, e.g., .95, c(.95,.99), al15() or al49(), etc.}

\item{Q}{a vector of quantiles (i.e., stresses)}
}
\value{
A vector containing the requested confidence limits
}
\description{
Compute confidence limits about probabilities and quantiles
}
