% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal2SampleDiffGNG.R
\name{Normal2SampleDiffGNG}
\alias{Normal2SampleDiffGNG}
\title{{Difference of means of two independent population with normal endpoint}}
\usage{
Normal2SampleDiffGNG(n.t, mu.t, sd.t, n.c, mu.c, sd.c, measure,
                     hypothesis = "greater than", threshold,
                     PP.cutoffGo = 0.60, PP.cutoffNoGo  = 0.60,
                     prior = "non-informative", mu0t, nu0t, kappa0t, var0t,
                     mu0c, nu0c, kappa0c, var0c, nsim = 10000, reps = 10000)
}
\arguments{
\item{n.t}{Sample size in treatment arm.}

\item{mu.t}{True mean of the treatment arm.}

\item{sd.t}{True sd of the treatment arm.}

\item{n.c}{Sample size in control arm.}

\item{mu.c}{True mean of the control arm.}

\item{sd.c}{True sd of the control arm.}

\item{measure}{Difference or ratio of means.}

\item{hypothesis}{"greater than" (Default): (PP(true mean>=k)>= c1 vs. PP(true mean>=k)< c2);
"less than": (PP(true mean<=k)>= c1 vs. PP(true mean<=k)< c2);
"between": (PP(k1<=true mean<=k2)>= c1 vs. PP(k1<=true mean<=k2)< c2).}

\item{threshold}{Target region cut-off (k); k is scaler for "greater than" and "less than" hypotheses; k is a vector of k1 and k2 for "between" hypothesis.}

\item{PP.cutoffGo}{Cut-off for probability of Go; Default is 0.60.}

\item{PP.cutoffNoGo}{Cut-off for probability of No-Go; Default is 0.60.}

\item{prior}{Prior choices: "non-informative" (Default), "informative".}

\item{mu0t}{Prior mean for treatment hyper-parameter if prior = "informative".}

\item{nu0t}{Prior sample size for treatment hyper-parameter if prior = "informative".}

\item{kappa0t}{Prior DF for treatment hyper-parameter if prior = "informative".}

\item{var0t}{Prior variance for treatment hyper-parameter if prior = "informative".}

\item{mu0c}{Prior mean for control hyper-parameter if prior = "informative".}

\item{nu0c}{Prior sample size for control hyper-parameter if prior = "informative".}

\item{kappa0c}{Prior DF for control hyper-parameter if prior = "informative".}

\item{var0c}{Prior variance for control hyper-parameter if prior = "informative".}

\item{nsim}{Number of simulated data sets; Default is 10000.}

\item{reps}{Number of repetition for samples from the posterior using simulation; Default is 10000.}

}

\value{
It returns a list of output including the inputs, posterior probability, probability of Go, and probability of No-Go.
}


\details{
To calculate the posterior probabilities with probabilities of Go and No-Go, the code follows the steps as below:
\itemize{
\item Generate \eqn{n.t} data from \eqn{Normal(mu.t, sd.t^2)}
\item Calculate sample mean \eqn{(\bar{y})} and sample variance \eqn{(s_y^2)}
\item Then draw samples from the posterior distribution using simulation
\itemize{
  \item Generate variances from posterior
  \itemize{
    \item Non-informative prior (See Appendix in Gelman et al. for scaled inverse chi square generation)
    \itemize{
      \item \eqn{s^2 \sim \chi^2(df = (n.t-1), ncp = 0); \sigma^2 = \frac{(n.t-1)*s_y^2}{s^2}}
      \item \eqn{prior.mu = \bar{y}; prior.sd = \sqrt\frac{\sigma^2}{n.t}; prior.df = n.t}
      }
    \item Informative prior
     \itemize{
      \item \eqn{var1 = (\frac{1}{(\nu_0 + n.t)}*(\nu_0*var0 + (n.t-1)*s_y^2 + (\frac{\kappa_0*n.t}{\kappa_0 + n.t}*(\bar{y} - \mu_0)^2)}, where \eqn{\mu_0, \nu_0, \kappa_0, var0}: Prior mean, sample size, degrees of freedom, variance for hyper-parameters
      \item \eqn{prior.mu = \frac{\kappa_0*\mu_0 + n.t*\bar{y}}{\kappa_0 + n.t}; prior.sd = \sqrt\frac{var1}{\kappa_0 + n.t}; prior.df = \nu_0 + n.t}
      }
      }
  \item Generate mean from posterior using normal distribution with mean = prior.mu and sd = prior.sd
    }
\item Similarly, calculate mean from posterior using the information from control arm
\item Calculate the posterior probabilities using the difference of treatment and control arm
\item Now, based on `nsim` iterations, calculate the probabilities of Go and No-Go using the pre-defined posterior probability cut-offs.
}
}
\description{
{Difference of means of two independent population with normal endpoint.}
}


\examples{
# If prior = "non-informative"
Normal2SampleGNG(n.t = 9, mu.t = 70, sd.t = 10, n.c = 3, mu.c = 40, sd.c = 15, measure = "difference", threshold = 30)

# If prior = "informative"
Normal2SampleGNG(n.t = 9, mu.t = 70, sd.t = 10, n.c = 3, mu.c = 40, sd.c = 15, measure = "difference", threshold = 30, prior = "informative", mu0t = 70, nu0t = 4, kappa0t = 7, var0t = 1, mu0t = 40, nu0t = 2, kappa0t = 2, var0t = 1)
}
