% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nphaseI.R
\name{nphaseI}
\alias{nphaseI}
\title{Title}
\usage{
nphaseI(dat0, mlo, mhi, sg, reso, about, titl, unit, ln)
}
\arguments{
\item{dat0}{}

\item{mlo}{}

\item{mhi}{}

\item{sg}{}

\item{reso}{}

\item{about}{}

\item{titl}{}

\item{unit}{}

\item{ln}{}
}
\description{
Title
}
