% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npI.R
\name{npI}
\alias{npI}
\title{Title}
\usage{
npI(mlo, mhi, sg, tmu, tsig, reso, ln, iseed, dat0 = data.frame(numeric(0)))
}
\arguments{
\item{mlo}{}

\item{mhi}{}

\item{sg}{}

\item{tmu}{}

\item{tsig}{}

\item{reso}{}

\item{ln}{}

\item{iseed}{}

\item{dat0}{}
}
\description{
Title
}

