% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyqrda.R
\name{nyqrda}
\alias{nyqrda}
\title{Title}
\usage{
nyqrda(dat, lgit = F, ln = F, xmin = -9999, xmax = 9999, conf = 0.95,
  small = F, response = 0, labx = "", laby = "PROBABILITY OF RESPONSE",
  maxitt = 10, eps = 1e-06, zee = 0)
}
\arguments{
\item{dat}{sensitivity test "w"}

\item{zee}{}
}
\description{
retrieve some basic information about a test w and the ML fit that was obtained
}
