% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherpoint.R
\name{otherpoint}
\alias{otherpoint}
\title{Title}
\usage{
otherpoint(rx, ry, muhat, levs0, con)
}
\arguments{
\item{rx}{}

\item{ry}{}

\item{muhat}{}

\item{levs0}{}

\item{con}{}
}
\description{
Title
}
