% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavdf.R
\name{pavdf}
\alias{pavdf}
\title{Title}
\usage{
pavdf(data.df, ln, plotit = F, response = 0, lineit = F,
  labx = "STIMULUS", laby = "PROBABILITY OF RESPONSE",
  titl = "PAV SOLUTION")
}
\arguments{
\item{data.df}{}

\item{ln}{}

\item{plotit}{}

\item{response}{}

\item{lineit}{}

\item{labx}{}

\item{laby}{}

\item{titl}{}
}
\description{
Title
}
