% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pdat2}
\alias{pdat2}
\title{MLE's of mu and sigma plot, console version}
\usage{
pdat2(dat)
}
\arguments{
\item{dat}{sensitivity test "w"}
}
\description{
MLE's of mu and sigma plot, console version
}
