% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pdat3}
\alias{pdat3}
\title{Response curve with data plot, console version}
\usage{
pdat3(dat)
}
\arguments{
\item{dat}{sensitivity test "w"}
}
\description{
Response curve with data plot, console version
}
