% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseI3.R
\name{phaseI3}
\alias{phaseI3}
\title{Title}
\usage{
phaseI3(d0, dat0, sg, reso, about, titl, unit, ln)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{sg}{}

\item{reso}{}

\item{about}{}

\item{titl}{}

\item{unit}{}

\item{ln}{}
}
\description{
Title
}
