% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseII.R
\name{phaseII}
\alias{phaseII}
\title{Title}
\usage{
phaseII(d0, dat0, n2, reso, about, titl, unit, ln)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{n2}{}

\item{reso}{}

\item{about}{}

\item{titl}{}

\item{unit}{}

\item{ln}{}
}
\description{
Title
}

