% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pI3.R
\name{pI3}
\alias{pI3}
\title{Title}
\usage{
pI3(d0, dat0, sg, tmu, tsig, reso, ln, iseed)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{sg}{}

\item{tmu}{}

\item{tsig}{}

\item{reso}{}

\item{ln}{}

\item{iseed}{}
}
\description{
Title
}
