% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pII.R
\name{pII}
\alias{pII}
\title{Title}
\usage{
pII(d0, dat0, tmu, tsig, n2, reso, ln, iseed)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{tmu}{}

\item{tsig}{}

\item{n2}{}

\item{reso}{}

\item{ln}{}

\item{iseed}{}
}
\description{
Title
}
