% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{plot.CLS_GONOGO_DEC_ALL}
\alias{plot.CLS_GONOGO_DEC_ALL}
\title{Print decision result}
\usage{
\method{plot}{CLS_GONOGO_DEC_ALL}(x, x_axis = c("y1", "delta"), hl_region = TRUE, ...)
}
\description{
Print decision result
}
