% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_utils_bin.R
\name{plot.CLS_GONOGO_POSTS_SINGLE}
\alias{plot.CLS_GONOGO_POSTS_SINGLE}
\title{Plot posterior of observed data}
\usage{
\method{plot}{CLS_GONOGO_POSTS_SINGLE}(x, option = c("effect", "rate"), ...)
}
\description{
Plot posterior of observed data
}
