% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn_decision.R
\name{print.CLS_OPC_BIN}
\alias{print.CLS_OPC_BIN}
\title{Print decision result}
\usage{
\method{print}{CLS_OPC_BIN}(x, option = c("long", "wide_n", "wide_rate"), ...)
}
\description{
Print decision result
}
