% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSdat2.R
\name{pSdat2}
\alias{pSdat2}
\title{MLE's of mu and sigma plot, simulation version}
\usage{
pSdat2(dat)
}
\arguments{
\item{dat}{sensitivity tests "w"}
}
\description{
MLE's of mu and sigma plot, simulation version
}
