% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSdat3.R
\name{pSdat3}
\alias{pSdat3}
\title{Response curve with data plot, simulation version}
\usage{
pSdat3(dat)
}
\arguments{
\item{dat}{sensitivity tests "w"}
}
\description{
Response curve with data plot, simulation version
}
