% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{ptest}
\alias{ptest}
\title{Generate graphical output from results obtained from gonogo or gonogoSim.}
\usage{
ptest(dat, plt, notitle = F)
}
\arguments{
\item{dat}{dataframe containing results from gonogo or gonogoSim}

\item{plt}{plot number to generate, must be between 1 and 8
1 = history plot
2 = MLE's of mu and sigma
3 = response curve, with data, Pooled Adjacent
    violators solution, and 100*conf 2-Sided
    confidence bounds
4 = a simple visual of the data
5 = joint likelihood ratio (LR) multi-confidence bounds
6 = joint & individual (LR) multi-confidence bounds
7 = joint and/or individual LR confidence bounds
8 = confidence bounds on probability (p) and quantile (q)
    computed via 3 methods: likelihood ratio (LR), Fisher
    matrix (FM) and General Linear Model (GLM)}

\item{notitle}{set to T to exclude the title, defaults to F}
}
\value{
Dataframe containing results.
}
\description{
Generate graphical output from results obtained from gonogo or gonogoSim.
}
