% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Quick reset of graphical parameters}
\usage{
reset()
}
\description{
\code{reset} resets the graphical parameters (\code{par}) arguments. Primarily
leveraged by \code{gonogo} plotting functions where parameters are adjusted
but are then reset after the plotting outputs are produced.
}
