% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.lims.R
\name{sp.lims}
\alias{sp.lims}
\title{Title}
\usage{
sp.lims(dat, conf, maxitt = 10, response = 1, rd = 6)
}
\arguments{
\item{dat}{}

\item{conf}{}

\item{maxitt}{}

\item{response}{}

\item{rd}{}
}
\description{
Title
}

