% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphaseIII.R
\name{sphaseIII}
\alias{sphaseIII}
\title{Title}
\usage{
sphaseIII(d0, dat0, n3, p, reso, about, titl, unit, ln, lam = 0)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{n3}{}

\item{p}{}

\item{reso}{}

\item{about}{}

\item{titl}{}

\item{unit}{}

\item{ln}{}

\item{lam}{}
}
\description{
Title
}

