% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spIIIsim.R
\name{spIIIsim}
\alias{spIIIsim}
\title{Title}
\usage{
spIIIsim(d0, dat0, tmu, tsig, n3, p, lam, reso, ln, iseed)
}
\arguments{
\item{d0}{}

\item{dat0}{}

\item{tmu}{}

\item{tsig}{}

\item{n3}{}

\item{p}{}

\item{lam}{}

\item{reso}{}

\item{ln}{}

\item{iseed}{}
}
\description{
Title
}

