% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uliknext.R
\name{uliknext}
\alias{uliknext}
\title{Title}
\usage{
uliknext(rx, ry, levs0, em1, es, em2)
}
\arguments{
\item{rx}{}

\item{ry}{}

\item{levs0}{}

\item{em1}{}

\item{es}{}

\item{em2}{}
}
\description{
Title
}

