% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbd.R
\name{unbd}
\alias{unbd}
\title{Title}
\usage{
unbd(rx, ry, levs, mh1, mh2, es, mlim)
}
\arguments{
\item{rx}{}

\item{ry}{}

\item{levs}{}

\item{mh1}{}

\item{mh2}{}

\item{es}{}

\item{mlim}{}
}
\description{
Title
}

