% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_user_response.R
\name{validate_user_response}
\alias{validate_user_response}
\title{Validates the user's response}
\usage{
validate_user_response(choice, choiceKeys)
}
\arguments{
\item{choice}{Character string specifying the chosen key}

\item{choiceKeys}{Character string specifying which keys are treated as the choice keys}
}
\value{

}
\description{
The function \emph{validate_user_response} validates the chosen key by comparing it to a list of predefined choice keys.
}
\examples{
validate_user_response(choice = "m", choiceKeys = c("m", "o"))

}
