% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wabout13.R
\name{wabout13}
\alias{wabout13}
\title{Title}
\usage{
wabout13(M, cmlo, cmhi, csg, p, n11, n12, n2, n3, lam, reso)
}
\arguments{
\item{M}{}

\item{cmlo}{}

\item{cmhi}{}

\item{csg}{}

\item{p}{}

\item{n11}{}

\item{n12}{}

\item{n2}{}

\item{n3}{}

\item{lam}{}

\item{reso}{}
}
\description{
Title
}
