% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{wxdat}
\alias{wxdat}
\title{Get a specified sample data set}
\usage{
wxdat(ic, plt = T)
}
\arguments{
\item{ic}{the number of the data set to return, must be between 1 and 27}

\item{plt}{set to F to suppress plotting the data set, defaults to T}
}
\value{
a dataframe containing the ith data set
}
\description{
Get a specified sample data set
}
