% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rt.R
\name{check_rt}
\alias{check_rt}
\title{Check the reaction times in your data}
\usage{
check_rt(data, ratio = 1/3)
}
\arguments{
\item{data}{The output from gonogo()}

\item{ratio}{The ratio of the distance between any two (sorted)
data points. For example, the default 1/3 means that
the user is notified when there is a distance between
any two sorted data points that is greater than one
third of the range of the data.}
}
\value{
Returns tips for processing the reaction times.
}
\description{
The \emph{check_rt} function helps spot unusual patterns in your
reaction time data by pointing out unusually long distances between
sorted observations.
}
\details{
For more information about how to handle reaction time data,
please see \href{https://doi.org/10.3389/fpsyg.2021.675558}{article}.
}
\examples{
check_rt(data, ratio = 1/4)
}
