% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_screen.R
\name{draw_screen}
\alias{draw_screen}
\title{Draws a screen}
\usage{
draw_screen(txt, cex = 1, col = "black")
}
\arguments{
\item{txt}{A character string specifying the text to be printed on the screen}

\item{cex}{A numeric value specifying the font size (default is 1)}

\item{col}{A character string specifying the color of the text (default is black)}
}
\value{
Draws a screen corresponding to the specified values
}
\description{
Draws a screen
}
\examples{
draw_screen(txt = "Hello!", cex = 2, col = "red")
}
