% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixw.R
\name{fixw}
\alias{fixw}
\title{Title}
\usage{
fixw(w, k = 1)
}
\arguments{
\item{w}{sensitivity test}

\item{k}{}
}
\description{
An error correction feature: Suppose an error was inadvertently entered at the nth previous console read.
The syntax to go back, fix and continue is: z=fixw(w,n); w=gonogo(newz=F)
}
