% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixw.R
\name{fixw}
\alias{fixw}
\title{fixw}
\usage{
fixw(w, k = 1)
}
\arguments{
\item{w}{A sensitivity test saved as a list produced by \code{gonogo}}

\item{k}{A positive integer to indicate the last \code{k} entries to fix}
}
\value{
The updated list of the sensitivity test
}
\description{
An error correction feature: Suppose an error was inadvertently entered at the nth previous console read.
The syntax to go back, fix and continue is: z=fixw(w,n); w=gonogo(newz=F)
}
