% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{gonogo}
\alias{gonogo}
\title{gonogo}
\usage{
gonogo(mlo = 0, mhi = 0, sg = 0, newz = T, reso = 0, ln = F,
  neyer = F)
}
\arguments{
\item{mlo}{The min value \eqn{\mu_min} to guess a reasonable range (\eqn{\mu_min, \mu_max}) for \eqn{\mu}}

\item{mhi}{The max value \eqn{\mu_max} to guess a reasonable range (\eqn{\mu_min, \mu_max}) for \eqn{\mu}}

\item{sg}{Guess for standard deviation \eqn{\sigma_g}, should satisfy: \eqn{\mu_max - \mu_min \ge 6*\sigma_g}}

\item{newz}{logical; if \code{TRUE} creates a new list of sensitivity tests,
otherwise it looks for a list, saved as \emph{z}, in the immediate environment. So, reserve \emph{z} for that use.}

\item{reso}{Specified resolution used in testing.
For example: a "reso=.125" option recommends test levels rounded to the nearest one eighth.}

\item{ln}{logical; if \code{TRUE} test in log(\emph{X}) units (Experimental)}

\item{neyer}{logical; if \code{TRUE}, Neyer test is performed, otherwise three-phase optimal design (3pod) test}
}
\value{
A list containing 14 named objects, which are:
\enumerate{

		\item \code{d0}: A data frame having column names of ("X", "Y","COUNT","RX","EX","TX","ID"),
		where the columns contain the sequence of: X's, Y's, 1's, Recommended X's, Exact X's,
		Transformed X's (used for log transformed data), and ID's (Identifying the test Phase), respectively

		\item \code{about}: A string containing \code{mlo}, \code{mhi}, \code{sg},
		      \code{n1}, \code{n2}, \code{n3}, \code{p}, \code{lambda}, and \code{reso}

		\item \code{title}: A string containing the user defined title

		\item \code{units}: A string containing the user defined units

		\item \code{en}: A vector of the lengths (\emph{n1},\emph{n2},\emph{n3}) for the 3 Phases (thus far)

		\item \code{p}: The numeric value of the user define \emph{p} (0 if not in Phase III)

		\item \code{reso}: The resolution (0 default for no rounding of recommended stresses)

		\item \code{n2n3}: An integer that the function \code{fixw} may use when navigating backwards

		\item \code{ln}: A logical (\code{TRUE} or \code{FALSE}), \code{TRUE} when method operates on log transformed stresses

		\item \code{init}: A vector containing the user defined initial values \code{mlo}, \code{mhi}, and \code{sg}

		\item \code{lam}: A numeric value of the user entry lambda (0 prior to Phase III)

		\item \code{neyer}: A logical (\code{TRUE} or \code{FALSE}), \code{TRUE} for Neyer Test, \code{FALSE} for 3pod (default)

		\item \code{savinit}: A saved version of \code{init}

		\item \code{jvec}: An i by 9 matrix, \eqn{0 \le I \le n3+1} documenting the Phase III calculation
}
}
\description{
\code{gonogo} consists of dynamic sequential algorithms for sensitivity testing (binary response information).
It is a modern sensitivity test tool to conduct a 3pod (default) test or a Neyer test. This three-phase
optimal design of sensitivity experiments is highly modular. It can be used to meet several test objectives,
from exploratory testing (Phase I), to refined reliability estimates (Phase III). The three-phase procedure
can be viewed as a trilogy of "search-estimate-approximate."

There are three versions of 3pod, console(one at a time, keyboard entry), batch (a vector of responses is looked for)
and one suited for simulation. This function is the console version.

A 3pod test is begun by: w=gonogo(mlo,mhi,sg)

A Neyer test is begun by: w=gonogo(mlo,mhi,sg,neyer=T)
}
