% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{gonogo}
\alias{gonogo}
\title{Run adaptive sensitivity tests in console or batch mode.}
\usage{
gonogo(
  mlo = 0,
  mhi = 0,
  sg = 0,
  newz = T,
  reso = 0,
  ln = F,
  test = 1,
  term1 = T,
  BL = NULL,
  Y = NULL,
  X = NULL
)
}
\arguments{
\item{mlo}{mu low}

\item{mhi}{mu high}

\item{sg}{sigma guess}

\item{newz}{set to F to restart using global z, defaults to T}

\item{reso}{resolution of stimulus}

\item{ln}{set to T to use log transform which keeps recommended stimulus positive, defaults to F}

\item{test}{number corresponding to test type
1 = 3pod
2 = Neyer
3 = Bruceton
4 = Langlie}

\item{term1}{set to F to stay in phase I2 until overlap found, defaults to T}

\item{BL}{vector of 3 integers, needed for Bruceton or Langlie tests only}

\item{Y}{vector of responses used for batch mode, all elements must be 0 or 1}

\item{X}{vector of stimulus values used for batch mode}
}
\value{
Dataframe containing results.
}
\description{
Run adaptive sensitivity tests in console or batch mode.
}
