% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{gonogo}
\alias{gonogo}
\title{Run adaptive sensitivity tests in console or batch mode}
\usage{
gonogo(
  mlo = 0,
  mhi = 0,
  sg = 0,
  newz = T,
  reso = 0,
  ln = F,
  test = 1,
  term1 = T,
  BL = NULL,
  Y = NULL,
  X = NULL
)
}
\arguments{
\item{mlo}{Guess for \ifelse{html}{\out{<i>&mu;</i><sub>min</sub>}}{\eqn{\mu_{min}}}, which
is the low end of the range where the 50/50 point is expected.}

\item{mhi}{Guess for \ifelse{html}{\out{<i>&mu;</i><sub>max</sub>}}{\eqn{\mu_{max}}}, which
is the high end of the range where the 50/50 point is expected.}

\item{sg}{Guess for \ifelse{html}{\out{<i>&sigma;</i><sub>g</sub>}}{\eqn{\sigma_{g}}}, which
is a guess of how quickly the probability of the output being 1 rises with the input
variable x. Unless this is known, a value of (mhi-mlo)/6 is recommended.}

\item{newz}{Set to F to restart using global z, defaults to T.}

\item{reso}{Resolution of input variable x, will be used for rounding
recommended test values. For example, if reso is set to 0.01, then
all test values will be rounded to the hundredths place.}

\item{ln}{Set to T to use log transform which keeps recommended stimulus positive,
defaults to F.}

\item{test}{Number corresponding to test type.
* 1 = 3pod
* 2 = Neyer
* 3 = Bruceton
* 4 = Langlie}

\item{term1}{Set to F to stay in phase I2 until overlap found, defaults to T.}

\item{BL}{Vector of 3 integers, needed for Bruceton or Langlie tests only.}

\item{Y}{Vector of responses used for batch mode, all elements must be 0 or 1.}

\item{X}{Vector of stimulus values used for batch mode.}
}
\value{
List containing results.
}
\description{
Run adaptive sensitivity tests in console or batch mode
}
