% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\name{gonogo}
\alias{gonogo}
\title{Run adaptive sensitivity tests in console or batch mode}
\usage{
gonogo(
  mlo = 0,
  mhi = 0,
  sg = 0,
  newz = T,
  reso = 0,
  ln = F,
  test = 1,
  term1 = T,
  BL = NULL,
  Y = NULL,
  X = NULL
)
}
\arguments{
\item{mlo}{Guess for low value of mu.}

\item{mhi}{Guess for high value of mu.}

\item{sg}{Guess for sigma.}

\item{newz}{Set to F to restart using global z, defaults to T.}

\item{reso}{Resolution of stimulus.}

\item{ln}{Set to T to use log transform which keeps recommended stimulus positive,
defaults to F.}

\item{test}{Number corresponding to test type.
* 1 = 3pod
* 2 = Neyer
* 3 = Bruceton
* 4 = Langlie}

\item{term1}{Set to F to stay in phase I2 until overlap found, defaults to T.}

\item{BL}{Vector of 3 integers, needed for Bruceton or Langlie tests only.}

\item{Y}{Vector of responses used for batch mode, all elements must be 0 or 1.}

\item{X}{Vector of stimulus values used for batch mode.}
}
\value{
List containing results.
}
\description{
Run adaptive sensitivity tests in console or batch mode
}
