% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R
\docType{package}
\name{gonogo}
\alias{gonogo}
\title{This package contains the Go-No Go Task.}
\description{
The package has x functions.
}
\section{Validate user's response}{

\emph{validate_user_response} validates the user's response.
}

