% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogo.R, R/main.R
\docType{package}
\name{gonogo}
\alias{gonogo}
\title{This package contains the Go-No Go Task.}
\usage{
gonogo(id, n_trial, stimuli = c("A", "X"), inter = 0.5, prb = c(0.75, 0.25))
}
\arguments{
\item{id}{Either a character string specifying the participant's name, or a numeric vector of length 1 specifying the participant's id number}

\item{n_trial}{Number of trials (numeric vector of length 1)}

\item{stimuli}{A character vector of length 2: first element specifies the 'go' element, second stimulus specifies the 'no go' stimuls (default is c("A", "X"))}

\item{inter}{Length of each trial in seconds (numeric vector of length 1) (default is 0.6)}

\item{prb}{A numeric vector of length 2 specifying the probability weights for the stimuli (default is c(0.5, 0.5))}
}
\value{
A dataframe consisting of \emph{n_trial} rows and six columns: \emph{id} (participant's name or numeric id),  \emph{response} (response key used on the trial),  \emph{correct} (1=correct, 0=incorrect),  \emph{SDT} (responses categorised according to Signal Detection Theory),  \emph{rt} (reaction time), and  \emph{stimulus} (the stimulus shown on the trial)
}
\description{
The package has x functions.
}
\section{Validate user's response}{

\emph{validate_user_response} validates the user's response.
}

\examples{
p1_data <- gonogo(id = "p1", n_trial = 50, stimuli = c("T", "S"), inter = 0.6, prb = c(0.7, 0.3))
}
