% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonogoSim.R
\name{gonogoSim}
\alias{gonogoSim}
\title{Title}
\usage{
gonogoSim(mlo, mhi, sg, n2, n3, p, lam, dm = 0, ds = 0, ln = F, plt = 0,
  neyer = F, iseed = -1, IIgo = T, M = 1)
}
\arguments{
\item{mlo}{}

\item{mhi}{}

\item{sg}{}

\item{n2}{}

\item{n3}{}

\item{p}{}

\item{lam}{}

\item{dm}{}

\item{ds}{}

\item{ln}{}

\item{plt}{}

\item{neyer}{}

\item{iseed}{}

\item{IIgo}{}

\item{M}{}
}
\description{
Title
}

