% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlrcb.R
\name{jlrcb}
\alias{jlrcb}
\title{Joint & Individual likelihood ratio confidence bounds plot}
\usage{
jlrcb(dat)
}
\arguments{
\item{dat}{sensitivity test "w"}
}
\description{
Joint & Individual likelihood ratio confidence bounds plot
}
