% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrmax.R
\name{lrmax}
\alias{lrmax}
\title{lrmax}
\usage{
lrmax(w, plt = F)
}
\arguments{
\item{w}{A sensitivity test saved as a list produced by \code{gonogo}}

\item{plt}{logical; if \code{TRUE} display a simple visual of the data}
}
\value{
This function returns several test data features:

the overlap category "one23" (1 for interval overlap; 2 for point overlap; and 3 for no overlap);
c1max and c2max, which indicate whether the LR joint confidence regions are bounded (conf1 < c1max, or conf2 < c2max);
and con = # of responses/total tested (which determines the axis of the joint LR confidence region).
conf1 and conf2 are related by: c1max=pchisq(qchisq(c2max,2),1).
}
\description{
Lists several important characteristics of the test data pertinent to likelihood ratio
confidence intervals and regions
}
