% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Plots Console Version}
\usage{
ptest(dat, plt)
}
\arguments{
\item{dat}{A sensitivity test saved as a list produced by \code{gonogo}}

\item{plt}{An integer between 1 and 8 to indicate which plot is wanted:
\enumerate{
        \item 1: History plot
        \item 2: MLE's of mu and sigma
        \item 3: Response curve (with data), the pooled adjacent violators (PAV) solution (an optimal step function solution),
        and 95\% 1-sided confidence bounds computed via \code{glm}
        \item 4: A simple visual of the data
        \item 5: Joint LR multi-confidence bounds
        \item 6: Joint & Individual LR multi-confidence bounds
        \item 7: Joint and/or individual LR confidence bounds
        \item 8: Confidence bounds on probability (p) and quantile (q) computed via 3 methods:
        Likelihood Ratio (LR), Fisher Matrix (FM) and General Linear Model (GLM)
}}
}
\description{
Plots Console Version
}
