% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrt.R
\name{rrt}
\alias{rrt}
\title{Calculates reaction time (in seconds) for pressing a specific key}
\usage{
rrt(choiceKeys = c(" "))
}
\arguments{
\item{choiceKeys}{A character vector specifying the key(s) for which reaction time is recorded}
}
\value{
Response (key pressed) and reaction time in seconds
}
\description{
Calculates reaction time (in seconds) for pressing a specific key
}
