% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update the Go-No Go data within the function}
\usage{
update_data(data, stimuli = stimuli)
}
\arguments{
\item{data}{The dataframe in the unedited format}

\item{stimuli}{The Go stimulus and No Go stimulus in a character vector of length 2, as specified in the gonogo() function}
}
\value{
The updated dataframe with six columns: \emph{n_trial*n_block} rows and seven columns: \emph{id} (participant's name or numeric id),  \emph{response} (response key used on the trial),  \emph{correct} (1=correct, 0=incorrect),  \emph{SDT} (responses categorized according to Signal Detection Theory),  \emph{rt} (reaction time in seconds),  \emph{stimulus} (the stimulus shown on the trial), and \emph{block} (the number of the block)
}
\description{
Update the Go-No Go data within the function
}
