% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wxdat.R
\name{wxdat}
\alias{wxdat}
\title{Data sets used to benchmark LR plots, particularly jlrbc and lrbc and cbs}
\usage{
wxdat(ic)
}
\arguments{
\item{ic}{for ic= 1,...,25 defines various data sets (in list format) considered while developing these LR graphics}
}
\description{
Data sets used to benchmark LR plots, particularly jlrbc and lrbc and cbs
}
